requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
audioop-lts==0.2.1
cachetools==5.5.2
certifi==2025.6.15
charset-normalizer==3.4.2
discord.py==2.5.2
dotenv==0.9.9
frozenlist==1.7.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-compute==1.31.0
googleapis-common-protos==1.70.0
grpcio==1.73.1
grpcio-status==1.73.1
idna==3.10
multidict==6.5.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
python-dotenv==1.1.1
requests==2.32.4
rsa==4.9.1
urllib3==2.5.0
yarl==1.20.1
